PC2/consumer/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
fastapi==0.104.1
uvicorn==0.24.0